Refactoring   - utils package for utility function like Hasher   - Separation of blocks and chain responsibility
